Rename gameview to game_view
