+++ src/view/game_view.rs
