+++ src/view/game_view/whole.rs
