+++ src/view/game_view/whole/finder.rs
