+++ src/view/game_view/whole/header.rs
